+++ txnprovider/txpool/pool_getblobs_test.go
@@ -19,8 +19,6 @@ package txpool
 import (
 	"context"
 	"math"
-	"runtime"
-	"sync"
 	"testing"
 
 	"github.com/holiman/uint256"
@@ -97,65 +95,6 @@ func addTestBlobTxn(tb testing.TB, pool *TxPool, ctx context.Context, senderByte
 		require.Equal(tb, txpoolcfg.Success, r, r.String())
 	}
 	return blobTxn.GetBlobHashes()
-}
-
-func BenchmarkGetBlobs(b *testing.B) {
-	const writerCriticalSectionWork = 2000
-
-	pool, ctx := newGetBlobsTestPool(b, 1)
-	realHashes := addTestBlobTxn(b, pool, ctx, 1, 0)
-
-	query := append([]common.Hash{}, realHashes...)
-	for i := 0; i < 4; i++ {
-		query = append(query, common.Hash{0xff, byte(i)})
-	}
-	got := pool.GetBlobs(query)
-	require.Len(b, got, len(query))
-	require.NotNil(b, got[0].Blob, "real blob hash must resolve")
-
-	b.Run("NoContention", func(b *testing.B) {
-		b.ReportAllocs()
-		for i := 0; i < b.N; i++ {
-			_ = pool.GetBlobs(query)
-		}
-	})
-
-	b.Run("WithPoolWriters", func(b *testing.B) {
-		stop := make(chan struct{})
-		var wg sync.WaitGroup
-		var writerCounter uint64
-		writers := max(2, runtime.GOMAXPROCS(0)-1)
-		for w := 0; w < writers; w++ {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				for {
-					select {
-					case <-stop:
-						return
-					default:
-					}
-					pool.lock.Lock()
-					for i := 0; i < writerCriticalSectionWork; i++ {
-						writerCounter++
-					}
-					pool.lock.Unlock()
-				}
-			}()
-		}
-
-		b.ReportAllocs()
-		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
-			_ = pool.GetBlobs(query)
-		}
-		b.StopTimer()
-		close(stop)
-		wg.Wait()
-		if writerCounter == 0 {
-			b.Fatal("writers did not run")
-		}
-	})
 }
 
 func TestGetBlobsRemovedAfterMined(t *testing.T) {
